web-llm-sql-csv/requirements.txt
flask
google-genai
sqlalchemy
pymssql
python-dotenv
pandas
openpyxl
gunicorn